requirements.txt
numpy == 1.18.1
pandas == 1.0.5
tensorflow == 2.3.1
python-dotenv==0.14.0
pymysql
click
protobuf == 3.20.3
mlflow == 2.10.1
jupyter
watermark == 2.4.3